react-router related files created and implemented
