+++ src/components/LandingPage.js
@@ -1,0 +1,19 @@
+import React from "react";
+
+const LandingPage = () => {
+  return (
+    <div
+      className="ui raised segment"
+      style={{
+        height: "100px",
+        width: "500px",
+        margin: "auto",
+        textAlign: "center"
+      }}
+    >
+      <h3 style={{ margin: "20px auto" }}>Please make a search...</h3>
+    </div>
+  );
+};
+
+export default LandingPage;
