+++ src/components/MovieCard.js
@@ -48,4 +48,3 @@ const mapStateToProps = state => ({
 });
 
 export default connect(mapStateToProps)(MovieCard);
-
